LayoutTests/http/tests/security/xssAuditor/block-does-not-leak-referrer.html

<!DOCTYPE html>
<html>
<head>
    <script src="/resources/js-test-pre.js"></script>
    <script>
        if (window.testRunner)
            testRunner.setXSSAuditorEnabled(true);
        window.jsTestIsAsync = true;

        function checkReferer() {
            window.frame = document.querySelector('iframe');
            shouldBeNull('frame.contentDocument');
            finishJSTest();
        }

        window.onload = function () {
            var i = document.querySelector('iframe');
            i.onload = checkReferer;
            i.src = 'http://localhost:8000/security/xssAuditor/resources/redir.php?url=http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?enable-full-block=1%26q=<script>alert(String.fromCharCode(0x58,0x53,0x53))<' + '/script>';
        };
    </script>
    <script src="/resources/js-test-post.js"></script>
</head>
<body>
    <iframe></iframe>
</body>
</html>
